Feat: 홈 화면의 뉴스 파넬 에러 및 로딩 상태 처리
