+++ src/components/home/organisms/NewsPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './NewsPanel';
